// packages/web-common/src/components/UI/Select/type.ts
export declare type Key = string | number;
export interface LabelValueType {
  key?: Key;
  value?: RawValueType;
  label?: React.ReactNode;
  isCacheable?: boolean;
}
export interface OptionCoreData {
  key?: Key;
  disabled?: boolean;
  value: Key;
  title?: string;
  className?: string;
  style?: React.CSSProperties;
  label?: React.ReactNode;
  /** @deprecated Only works when use `children` as option data */
  children?: React.ReactNode;
}
export interface OptionData extends OptionCoreData {
  /** Save for customize data */
  [prop: string]: any;
}